move: Emailmessage to infra
